# backend/src/main/java/io/github/alexeyzarechnev/autoservice/controller/PartsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.github.alexeyzarechnev.autoservice.controller;

import io.github.alexeyzarechnev.autoservice.dto.PartDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-24T20:18:11.476785+03:00[Europe/Moscow]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "Parts", description = "Tag for operations with parts in storage")
public interface PartsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /v1/storage/parts : Create a new part
     *
     * @param partDTO  (required)
     * @return Part successfully created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "createPart",
        summary = "Create a new part",
        tags = { "Parts" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Part successfully created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PartDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/storage/parts",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PartDTO> createPart(
        @Parameter(name = "PartDTO", description = "", required = true) @Valid @RequestBody PartDTO partDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"articleNumber\" : 6, \"price\" : 5, \"name\" : \"name\", \"remains\" : 1, \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "\"invalid body\"";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /v1/storage/parts/{id} : Delete a part from the catalog
     *
     * @param id  (required)
     * @return Part successfully deleted (status code 204)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "deletePart",
        summary = "Delete a part from the catalog",
        tags = { "Parts" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Part successfully deleted"),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/storage/parts/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deletePart(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "\"resource not found\"";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /v1/storage/parts/{id} : Get part by ID
     *
     * @param id  (required)
     * @return Part found (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getPartById",
        summary = "Get part by ID",
        tags = { "Parts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Part found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PartDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/storage/parts/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PartDTO> getPartById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"articleNumber\" : 6, \"price\" : 5, \"name\" : \"name\", \"remains\" : 1, \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "\"resource not found\"";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /v1/storage/parts/{id} : Update part
     *
     * @param id  (required)
     * @param partDTO  (required)
     * @return Part successfully updated (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "updatePart",
        summary = "Update part",
        tags = { "Parts" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Part successfully updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PartDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/storage/parts/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PartDTO> updatePart(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "PartDTO", description = "", required = true) @Valid @RequestBody PartDTO partDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"articleNumber\" : 6, \"price\" : 5, \"name\" : \"name\", \"remains\" : 1, \"id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "\"resource not found\"";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}