# backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: autoservice
  description:
  version: v1
  contact:
    name: AlexeyZarechnev

components:
  schemas:
    PartDTO:
      type: object
      description: Part representation
      properties:
        id:
          type: integer
          format: int64
          description: Unique part identifier
        name:
          type: string
          maxLength: 255
          description: Part name
        articleNumber:
          type: integer
          format: int64
          description: Article number
        remains:
          type: integer
          format: int16
          description: Available stock quantity
        price:
          type: integer
          format: int32
          description: Part price
      required:
        - name
        - articleNumber
        - remains
        - price

  responses:
    400BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: string
            example: invalid body

    404NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: string
            example: resource not found

tags:
  - name: Ping
    description: Tag for ping operation
  - name: Parts
    description: Tag for operations with parts in storage

paths:
  /v1/ping:
    get:
      tags: [Ping]
      summary: "Check if server is alive"
      description: "Returns pong as answer"
      operationId: pong
      responses:
        200:
          description: pong
          content:
            application/json:
              schema:
                type: string
                example: "pong"

  /v1/storage/parts:
    post:
      tags: [Parts]
      summary: Create a new part
      operationId: createPart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartDTO'
      responses:
        '201':
          description: Part successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartDTO'
        '400':
          $ref: '#/components/responses/400BadRequest'

  /v1/storage/parts/{id}:
    get:
      tags: [Parts]
      summary: Get part by ID
      operationId: getPartById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Part found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartDTO'
        '404':
          $ref: '#/components/responses/404NotFound'

    post:
      tags: [Parts]
      summary: Update part
      operationId: updatePart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartDTO'
      responses:
        '200':
          description: Part successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartDTO'
        '404':
          $ref: '#/components/responses/404NotFound'

    delete:
      tags: [Parts]
      summary: Delete a part from the catalog
      operationId: deletePart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Part successfully deleted
        '404':
          $ref: '#/components/responses/404NotFound'

        